+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Add PUT support to APId for the various resource types.
 - Added flags to disable the agent's API and Socket listeners
 - Made Changelog examples in CONTRIBUTING.md more obvious
+- Added cli support for setting environment variables in mutators and handlers.
 
 ### Changed
 - Changed the maximum number of open file descriptors on a system to from 1024